perf: split rust using feature cfg
