+++ kursal-cli/Cargo.toml
@@ -14,16 +14,28 @@ workspace = true
 name = "kursal-relay"
 path = "src/main.rs"
 
+[features]
+default = ["tui"]
+tui = ["dep:ratatui"]
+
 [dependencies]
 kursal-core = { path = "../kursal-core", default-features = false }
 
-clap = { workspace = true, features = ["derive"] }
-toml = { workspace = true }
-serde = { workspace = true, features = ["derive"] }
-tokio = { workspace = true, features = ["signal"]}
-log = { workspace = true }
-libp2p = { workspace = true }
-anyhow = { workspace = true }
+# Async
+tokio = { workspace = true, features = ["signal"] }
 
+# Serialization
+serde = { workspace = true, features = ["derive"] }
+toml = { workspace = true }
+
+# Errors & logging
+anyhow = { workspace = true }
+log = { workspace = true }
+
+# Network
+libp2p = { workspace = true }
 axum = { workspace = true }
-ratatui = { workspace = true }
+
+# CLI
+clap = { workspace = true, features = ["derive"] }
+ratatui = { workspace = true, optional = true }
