+++ kursal-tauri/src-tauri/Cargo.toml
@@ -14,43 +14,51 @@ workspace = true
 name = "kursal_tauri_lib"
 crate-type = ["staticlib", "cdylib", "rlib"]
 
-[build-dependencies]
-tauri-build = { workspace = true, features = [] }
-
 [dependencies]
-kursal-core = { path = "../../kursal-core" }
-kursal-cli = { path = "../../kursal-cli" }
+kursal-core = { path = "../../kursal-core", features = ["apiserver"] }
+kursal-cli = { path = "../../kursal-cli", default-features = false }
 
+# Tauri
 tauri = { workspace = true, features = [
     "protocol-asset",
     "devtools",
     "image-png",
     "tray-icon",
 ] }
+tauri-plugin-clipboard-manager = { workspace = true }
+tauri-plugin-deep-link = { workspace = true }
 tauri-plugin-dialog = { workspace = true }
 tauri-plugin-fs = { workspace = true }
+tauri-plugin-notification = { workspace = true }
 tauri-plugin-opener = { workspace = true }
-serde = { workspace = true, features = ["derive"] }
-clap = { workspace = true, features = ["derive"] }
-serde_json = { workspace = true }
+tauri-plugin-os = { workspace = true }
+tauri-plugin-sharekit = { workspace = true }
 
-anyhow = { workspace = true }
+# Async
 tokio = { workspace = true }
-hex = { workspace = true }
-base64 = { workspace = true }
-rayon = { workspace = true }
 async-trait = { workspace = true }
 
+# Serialization
+serde = { workspace = true, features = ["derive"] }
+serde_json = { workspace = true }
+hex = { workspace = true }
+base64 = { workspace = true }
+
+# Errors & logging
+anyhow = { workspace = true }
 log = { workspace = true }
 sentry = { workspace = true }
 sentry-panic = { workspace = true }
-tauri-plugin-clipboard-manager = { workspace = true }
-tauri-plugin-os = { workspace = true }
-tauri-plugin-deep-link = { workspace = true }
-tauri-plugin-notification = { workspace = true }
-tauri-plugin-sharekit = { workspace = true }
+
+# Misc
+clap = { workspace = true, features = ["derive"] }
+rayon = { workspace = true }
+
+[build-dependencies]
+tauri-build = { workspace = true, features = [] }
 
 [target.'cfg(not(any(target_os = "android", target_os = "ios")))'.dependencies]
+kursal-cli = { path = "../../kursal-cli", default-features = false, features = ["tui"] }
 tauri-plugin-autostart = { workspace = true }
 tauri-plugin-single-instance = { workspace = true }
 tauri-plugin-updater = { workspace = true }
@@ -62,8 +70,12 @@ tauri-plugin-barcode-scanner = { workspace = true }
 tauri-plugin-biometric = { workspace = true }
 tauri-plugin-haptics = { workspace = true }
 
-[target.'cfg(target_os = "android")'.dependencies]
-jni = "0.19"
+[target.'cfg(target_os = "macos")'.dependencies]
+objc2-foundation = { workspace = true, features = [
+    "std",
+    "NSString",
+    "NSUserDefaults",
+] }
 
 [target.'cfg(target_os = "ios")'.dependencies]
 objc2-foundation = { workspace = true, features = [
@@ -73,12 +85,8 @@ objc2-foundation = { workspace = true, features = [
     "NSURL",
 ] }
 
-[target.'cfg(target_os = "macos")'.dependencies]
-objc2-foundation = { workspace = true, features = [
-    "std",
-    "NSString",
-    "NSUserDefaults",
-] }
+[target.'cfg(target_os = "android")'.dependencies]
+jni = "0.19"
 
 [target.'cfg(target_os = "windows")'.dependencies]
 webview2-com = { workspace = true }
